botones

+++ app/app.py
@@ -110,10 +110,10 @@ def boton4():
         GPIO.cleanup()
         while(GPIO.input(In_1)):
             GPIO.output(Out_2, GPIO.HIGH)
-            sleep(tiempo1*1000)
+            sleep(int(tiempo1*1000))
             GPIO.output(Out_1, GPIO.HIGH)
         GPIO.output(Out_1, GPIO.LOW)
-        sleep(tiempo2*1000)
+        sleep(int(tiempo2*1000))
         GPIO.output(Out_2,GPIO.LOW)
         if GPIO.input(In_2):
             manual=0
